Clean up examples and sort them alphabetically
Examples are now sorted alphabetically by their label.
Improved labels of various examples to better reflect what they contain.

Removed examples:
- day 1 to day 6: Contain various elements covered in other examples.
- 1169: Covers specific tester bug, reusable tester is covered by core unit tests.
- 1220: Old issue from 2.2
- 1253: Old issue from 2.2, multiple other oneOf examples
- 1254: Integrated in enum.ts
- 1645: Performance / debouncing can be tested with the huge example
- 1996: oneOf part is covered by oneOfArray.ts, anyOf part was integrated in anyOf.ts
- booleanToggle: Covered in control-options.ts

Renamed examples:
- 1779.ts -> list-with-detail-primitives.ts
- stepper.ts -> categorization-stepper.ts
- steppershownav.ts -> categorization-stepper-nav-buttons.ts
- dynamic.ts -> generate-dynamic.ts

Part of #1706
