--- packages/examples/src/examples/1996.ts
@@ -1,102 +1,0 @@
-/*
-  The MIT License
-  
-  Copyright (c) 2022 EclipseSource
-  https://github.com/eclipsesource/jsonforms
-  
-  Permission is hereby granted, free of charge, to any person obtaining a copy
-  of this software and associated documentation files (the 'Software'), to deal
-  in the Software without restriction, including without limitation the rights
-  to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-  copies of the Software, and to permit persons to whom the Software is
-  furnished to do so, subject to the following conditions:
-  
-  The above copyright notice and this permission notice shall be included in
-  all copies or substantial portions of the Software.
-  
-  THE SOFTWARE IS PROVIDED 'AS IS', WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-  IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-  FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-  AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-  LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
-  THE SOFTWARE.
-*/
-import { registerExamples } from '../register';
-import { ControlElement } from '@jsonforms/core';
-
-export const schemaAnyOf = {
-  type: "array",
-  items: {
-    anyOf: [
-      {
-        type: "object",
-        properties: {
-          value: {
-            type: "string"
-          }
-        },
-        required: ["value"]
-      },
-      {
-        type: "object",
-        properties: {
-          value: {
-            type: "number"
-          }
-        },
-        required: ["value"]
-      }
-    ]
-  }
-};
-
-export const schemaOneOf = {
-  type: "array",
-  items: {
-    oneOf: [
-      {
-        type: "object",
-        properties: {
-          value: {
-            type: "string"
-          }
-        },
-        required: ["value"]
-      },
-      {
-        type: "object",
-        properties: {
-          value: {
-            type: "number"
-          }
-        },
-        required: ["value"]
-      }
-    ]
-  }
-};
-
-export const uischema: ControlElement = {
-  type: 'Control',
-  scope: '#'
-};
-
-export const data: any[] = [];
-
-registerExamples([
-  {
-    name: '1996_anyOf',
-    label: 'Issue 1996 - array with anyOf items',
-    data,
-    schema: schemaAnyOf,
-    uischema
-  },
-  {
-    name: '1996_oneOf',
-    label: 'Issue 1996 - array with oneOf items',
-    data,
-    schema: schemaOneOf,
-    uischema
-  }
-]);
