+++ packages/examples/src/examples/categorization-stepper.ts
