+++ packages/examples/src/examples/generate-dynamic.ts
@@ -1,19 +1,19 @@
 /*
   The MIT License
-
+  
   Copyright (c) 2017-2019 EclipseSource Munich
   https://github.com/eclipsesource/jsonforms
-
+  
   Permission is hereby granted, free of charge, to any person obtaining a copy
   of this software and associated documentation files (the "Software"), to deal
   in the Software without restriction, including without limitation the rights
   to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
   copies of the Software, and to permit persons to whom the Software is
   furnished to do so, subject to the following conditions:
-
+  
   The above copyright notice and this permission notice shall be included in
   all copies or substantial portions of the Software.
-
+  
   THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
   IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
   FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
@@ -22,41 +22,28 @@
   OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
   THE SOFTWARE.
 */
+import { StateProps } from '../example';
 import { registerExamples } from '../register';
-import { data as day1Data, schema as day1Schema } from './day1';
 
-export const schema = day1Schema;
-
-export const uischema = {
-  type: 'VerticalLayout',
-  elements: [
-    {
-      type: 'Control',
-      scope: '#/properties/name'
-    },
-    {
-      type: 'Control',
-      label: false,
-      scope: '#/properties/done'
-    },
-    {
-      type: 'Control',
-      scope: '#/properties/description',
-      options: {
-        multi: true
+const actions = [
+  {
+    'label': 'Change data',
+    'apply': (props: StateProps) => {
+      return {
+        ...props,
+        data: { id: 'aaa' }
       }
     }
-  ]
-};
-
-export const data = day1Data;
+  },
+]
 
 registerExamples([
   {
-    name: 'day2',
-    label: 'Day 2',
-    data,
-    schema,
-    uischema
+    name: 'dynamic',
+    label: 'Generate both schemas - Dynamic data change',
+    schema: undefined,
+    uischema: undefined,
+    data: { name: 'bla' },
+    actions
   }
 ]);
