+++ packages/examples/src/register.ts
@@ -30,5 +30,8 @@ export const registerExamples = (examples: ExampleDescription[]): void => {
   examples.forEach(example => (knownExamples[example.name] = example));
 };
 
-export const getExamples: () => ExampleDescription[] = () =>
-  Object.keys(knownExamples).map(key => knownExamples[key]);
+export const getExamples: () => ExampleDescription[] = () => {
+  const examples = Object.keys(knownExamples).map(key => knownExamples[key]);
+  examples.sort((a, b) => a.label.localeCompare(b.label));
+  return examples;
+}
